Try moving the SensorDict
